+++ src/vm.c
@@ -259,7 +259,7 @@ static InterpretResult run(void) {
                 break;
             }
             case OP_JMP: {
-                size_t label = READ2();
+                size_t label = (size_t) READ2();
                 int32_t offset = vm.chunk->labels[label];
 
                 if (offset < 0) {
@@ -273,7 +273,7 @@ static InterpretResult run(void) {
                 uint8_t reg = READ1();
                 int32_t val1 = AS_DWORD(vm.registers[reg]);
                 int32_t val2 = READ4();
-                size_t label = READ2();
+                size_t label = (size_t) READ2();
 
                 if (val1 == val2) {
                     int32_t offset = vm.chunk->labels[label];
